LayoutTests/http/tests/w3c/webperf/submission/Google/resource-timing/html/test_resource_frame_initiator_type.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title>Resource Timing frame initiator type</title>
<link rel="author" title="Google" href="http://www.google.com/" />
<link rel="help" href="http://www.w3.org/TR/resource-timing/#dom-performanceresourcetiming-initiatortype"/>
<script src="/w3c/resources/testharness.js"></script>
<script src="/w3c/resources/testharnessreport.js"></script>
<script src="/w3c/webperf/resources/webperftestharness.js"></script>
<script src="/w3c/webperf/resources/webperftestharnessextension.js"></script>
<script>
setup({explicit_done: true});

function onload_test() {
    var context = new PerformanceContext(iframe.contentWindow.performance);
    var entries = context.getEntries();

    var expected_entries = {
        '/w3c/webperf/resources/blank_page_green.htm': 'frame',
    };

    test_resource_entries(entries, expected_entries);
    done();
}
</script>
</head>
<body>
<h1>Description</h1>
<p>This test validates that the frame initiator type is represented.</p>
<div id="log"></div>
<iframe id="frameContext" onload="onload_test();" src="/w3c/webperf/resources/green_frame.htm" style="width: 250px; height: 250px;"></iframe>
</body>
</html>
