third_party/WebKit/LayoutTests/http/tests/w3c/webperf/submission/Google/resource-timing/html/test_resource_frame_initiator_type-expected.txt

Description

This test validates that the frame initiator type is represented.



PASS http://127.0.0.1:8000/w3c/webperf/resources/blank_page_green.htm is expected to have initiatorType frame 
